@@ -4,7 +4,6 @@
   import Icon from '$lib/components/elements/icon.svelte';
   import { TUNABLES } from '$lib/utils/tunables';
   import { mdiEyeOffOutline } from '@mdi/js';
-  import { onMount } from 'svelte';
   import { fade } from 'svelte/transition';
 
   interface Props {
@@ -37,7 +36,6 @@
     circle = false,
     hidden = false,
     border = false,
-    preload = true,
     hiddenIconClass = 'text-white',
     onComplete = undefined,
   }: Props = $props();
@@ -49,8 +47,6 @@
   let loaded = $state(false);
   let errored = $state(false);
 
-  let img = $state<HTMLImageElement>();
-
   const setLoaded = () => {
     loaded = true;
     onComplete?.();
@@ -59,11 +55,13 @@
     errored = true;
     onComplete?.();
   };
-  onMount(() => {
-    if (img?.complete) {
-      setLoaded();
+
+  function mount(elem: HTMLImageElement) {
+    if (elem.complete) {
+      loaded = true;
+      onComplete?.();
     }
-  });
+  }
 
   let optionalClasses = $derived(
     [
@@ -82,10 +80,9 @@
   <BrokenAsset class={optionalClasses} width={widthStyle} height={heightStyle} />
 {:else}
   <img
-    bind:this={img}
+    use:mount
     onload={setLoaded}
     onerror={setErrored}
-    loading={preload ? 'eager' : 'lazy'}
     style:width={widthStyle}
     style:height={heightStyle}
     style:filter={hidden ? 'grayscale(50%)' : 'none'}
