@@ -1,23 +1,201 @@
 import { locale } from '$lib/stores/preferences.store';
 import { getKey } from '$lib/utils';
-import { AssetGridTaskManager } from '$lib/utils/asset-store-task-manager';
-import { generateId } from '$lib/utils/generate-id';
-import { type getJustifiedLayoutFromAssetsFunction } from '$lib/utils/layout-utils';
-import type { AssetGridRouteSearchParams } from '$lib/utils/navigation';
-import { fromLocalDateTime, splitBucketIntoDateGroups, type DateGroup } from '$lib/utils/timeline-util';
-import { TimeBucketSize, getAssetInfo, getTimeBucket, getTimeBuckets, type AssetResponseDto } from '@immich/sdk';
-import { throttle } from 'lodash-es';
+import { CancellableTask } from '$lib/utils/cancellable-task';
+import {
+  getJustifiedLayoutFromAssets,
+  getPosition,
+  type CommonLayoutOptions,
+  type CommonPosition,
+} from '$lib/utils/layout-utils';
+import { formatDateGroupTitle, fromLocalDateTime } from '$lib/utils/timeline-util';
+import { TUNABLES } from '$lib/utils/tunables';
+import { getAssetInfo, getTimeBucket, getTimeBuckets, TimeBucketSize, type AssetResponseDto } from '@immich/sdk';
+import { debounce, isEqual, throttle } from 'lodash-es';
 import { DateTime } from 'luxon';
 import { t } from 'svelte-i18n';
+
 import { SvelteSet } from 'svelte/reactivity';
 import { get, writable, type Unsubscriber } from 'svelte/store';
 import { handleError } from '../utils/handle-error';
 import { websocketEvents } from './websocket';
 
-let getJustifiedLayoutFromAssets: getJustifiedLayoutFromAssetsFunction;
+const {
+  TIMELINE: { INTERSECTION_EXPAND_TOP, INTERSECTION_EXPAND_BOTTOM },
+} = TUNABLES;
+
+const THUMBNAIL_HEIGHT = 235;
+const GAP = 12;
+const HEADER = 49; //(1.5rem)
 
 type AssetApiGetTimeBucketsRequest = Parameters<typeof getTimeBuckets>[0];
-export type AssetStoreOptions = Omit<AssetApiGetTimeBucketsRequest, 'size'>;
+export type AssetStoreOptions = Omit<AssetApiGetTimeBucketsRequest, 'size'> & {
+  timelineAlbumId?: string;
+  deferInit?: boolean;
+};
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+function updateObject(target: any, source: any): boolean {
+  if (!target) {
+    return false;
+  }
+  let updated = false;
+  for (const key in source) {
+    // eslint-disable-next-line no-prototype-builtins
+    if (!source.hasOwnProperty(key)) {
+      continue;
+    }
+    if (typeof target[key] === 'object') {
+      updated = updated || updateObject(target[key], source[key]);
+    } else {
+      // Otherwise, directly copy the value
+      if (target[key] !== source[key]) {
+        target[key] = source[key];
+        updated = true;
+      }
+    }
+  }
+  return updated;
+}
+
+class IntersectingAsset {
+  // --- public ---
+  readonly #group: AssetDateGroup;
+
+  intersecting = $derived.by(() => {
+    if (!this.position) {
+      return false;
+    }
+
+    const store = this.#group.bucket.store;
+    const topWindow = store.visibleWindow.top + HEADER - INTERSECTION_EXPAND_TOP;
+    const bottomWindow = store.visibleWindow.bottom + HEADER + INTERSECTION_EXPAND_BOTTOM;
+    const positionTop = this.#group.absoluteDateGroupTop + this.position.top;
+    const positionBottom = positionTop + this.position.height;
+
+    const intersecting =
+      (positionTop >= topWindow && positionTop < bottomWindow) ||
+      (positionBottom >= topWindow && positionBottom < bottomWindow) ||
+      (positionTop < topWindow && positionBottom >= bottomWindow);
+    return intersecting;
+  });
+
+  position: CommonPosition | undefined = $state();
+  asset: AssetResponseDto | undefined = $state();
+  id: string = $derived.by(() => this.asset!.id);
+
+  constructor(group: AssetDateGroup, asset: AssetResponseDto) {
+    this.#group = group;
+    this.asset = asset;
+  }
+}
+type AssetOperation = (asset: AssetResponseDto) => { remove: boolean };
+
+type MoveAsset = { asset: AssetResponseDto; year: number; month: number };
+export class AssetDateGroup {
+  // --- public
+  readonly bucket: AssetBucket;
+  readonly index: number;
+  readonly date: DateTime;
+  readonly dayOfMonth: number;
+  intersetingAssets: IntersectingAsset[] = $state([]);
+  dodo: IntersectingAsset[] = $state([]);
+
+  height = $state(0);
+  width = $state(0);
+  intersecting = $derived.by(() => this.intersetingAssets.some((asset) => asset.intersecting));
+
+  // --- private
+  top: number = $state(0);
+  left: number = $state(0);
+  row = $state(0);
+  col = $state(0);
+
+  constructor(bucket: AssetBucket, index: number, date: DateTime, dayOfMonth: number) {
+    this.index = index;
+    this.bucket = bucket;
+    this.date = date;
+    this.dayOfMonth = dayOfMonth;
+  }
+
+  sortAssets() {
+    this.intersetingAssets.sort((a, b) => {
+      const aDate = DateTime.fromISO(a.asset!.fileCreatedAt).toUTC();
+      const bDate = DateTime.fromISO(b.asset!.fileCreatedAt).toUTC();
+      return bDate.diff(aDate).milliseconds;
+    });
+  }
+
+  getFirstAsset() {
+    return this.intersetingAssets[0]?.asset;
+  }
+  getRandomAsset() {
+    const random = Math.floor(Math.random() * this.intersetingAssets.length);
+    return this.intersetingAssets[random];
+  }
+
+  getAssets() {
+    return this.intersetingAssets.map((intersetingAsset) => intersetingAsset.asset!);
+  }
+
+  runAssetOperation(ids: Set<string>, operation: AssetOperation) {
+    if (ids.size === 0) {
+      return {
+        moveAssets: [] as MoveAsset[],
+        processedIds: new Set<string>(),
+        unprocessedIds: ids,
+        changedGeometry: false,
+      };
+    }
+    const unprocessedIds = new Set<string>(ids);
+    const processedIds = new Set<string>();
+    const moveAssets: MoveAsset[] = [];
+    let changedGeometry = false;
+    for (const assetId of unprocessedIds) {
+      const index = this.intersetingAssets.findIndex((ia) => ia.id == assetId);
+      if (index !== -1) {
+        const asset = this.intersetingAssets[index].asset!;
+        const oldTime = asset.localDateTime;
+        let { remove } = operation(asset);
+        const newTime = asset.localDateTime;
+        if (oldTime !== newTime) {
+          const utc = DateTime.fromISO(asset.localDateTime).toUTC().startOf('month');
+          const year = utc.get('year');
+          const month = utc.get('month');
+          if (this.bucket.year !== year || this.bucket.month !== month) {
+            remove = true;
+            moveAssets.push({ asset, year, month });
+          }
+        }
+        unprocessedIds.delete(assetId);
+        processedIds.add(assetId);
+        if (remove || this.bucket.store.isExcluded(asset)) {
+          this.intersetingAssets.splice(index, 1);
+          changedGeometry = true;
+        }
+      }
+    }
+    return { moveAssets, processedIds, unprocessedIds, changedGeometry };
+  }
+
+  layout(options: CommonLayoutOptions) {
+    const assets = this.intersetingAssets.map((intersetingAsset) => intersetingAsset.asset!);
+    const geometry = getJustifiedLayoutFromAssets(assets, options);
+    this.width = geometry.containerWidth;
+    this.height = assets.length === 0 ? 0 : geometry.containerHeight;
+    for (let i = 0; i < this.intersetingAssets.length; i++) {
+      const position = getPosition(geometry, i);
+      this.intersetingAssets[i].position = position;
+    }
+  }
+
+  get absoluteDateGroupTop() {
+    return this.bucket.top + this.top;
+  }
+
+  get groupTitle() {
+    return formatDateGroupTitle(this.date);
+  }
+}
 
 export interface Viewport {
   width: number;
@@ -28,111 +206,275 @@ export type ViewportXY = Viewport & {
   y: number;
 };
 
-interface AssetLookup {
-  bucket: AssetBucket;
-  bucketIndex: number;
-  assetIndex: number;
-}
-
 export class AssetBucket {
-  store!: AssetStore;
-  bucketDate: string = $state('');
+  // --- public ---
+  #intersecting: boolean = $state(false);
+  isLoaded: boolean = $state(false);
+  dateGroups: AssetDateGroup[] = $state([]);
+  readonly store: AssetStore;
+
+  // --- private ---
   /**
    * The DOM height of the bucket in pixel
    * This value is first estimated by the number of asset and later is corrected as the user scroll
    * Do not derive this height, it is important for it to be updated at specific times, so that
    * calculateing a delta between estimated and actual (when measured) is correct.
    */
-  bucketHeight: number = $state(0);
-  isBucketHeightActual: boolean = $state(false);
-  bucketDateFormattted!: string;
-  bucketCount: number = $derived.by(() => (this.isLoaded ? this.assets.length : this.initialCount));
-  initialCount: number = 0;
-  assets: AssetResponseDto[] = $state([]);
-  dateGroups: DateGroup[] = $state([]);
-  cancelToken: AbortController | undefined = $state();
-  /**
-   * Prevent this asset's load from being canceled; i.e. to force load of offscreen asset.
-   */
-  isPreventCancel: boolean = $state(false);
-  /**
-   * A promise that resolves once the bucket is loaded, and rejects if bucket is canceled.
-   */
-  complete!: Promise<void>;
-  loading: boolean = $state(false);
-  isLoaded: boolean = $state(false);
-  intersecting: boolean = $state(false);
-  measured: boolean = $state(false);
-  measuredPromise!: Promise<void>;
+  #bucketHeight: number = $state(0);
+  #top: number = $state(0);
+  #initialCount: number = 0;
 
-  constructor(props: Partial<AssetBucket> & { store: AssetStore; bucketDate: string }) {
-    Object.assign(this, props);
-    this.init();
+  // --- should be private, but is used by AssetStore ---
+
+  bucketCount: number = $derived(
+    this.isLoaded
+      ? this.dateGroups.reduce((accumulator, g) => accumulator + g.intersetingAssets.length, 0)
+      : this.#initialCount,
+  );
+  loader: CancellableTask | undefined;
+  isBucketHeightActual: boolean = $state(false);
+
+  readonly bucketDateFormatted: string;
+  readonly bucketDate: string;
+  readonly month: number;
+  readonly year: number;
+
+  constructor(store: AssetStore, utcDate: DateTime, initialCount: number) {
+    this.store = store;
+    this.#initialCount = initialCount;
+
+    const year = utcDate.get('year');
+    const month = utcDate.get('month');
+    const bucketDateFormatted = utcDate.toJSDate().toLocaleString(get(locale), {
+      month: 'short',
+      year: 'numeric',
+      timeZone: 'UTC',
+    });
+    this.bucketDate = utcDate.toISO()!.toString();
+    this.bucketDateFormatted = bucketDateFormatted;
+    this.month = month;
+    this.year = year;
+
+    this.loader = new CancellableTask(
+      () => {
+        this.isLoaded = true;
+      },
+      () => {
+        this.isLoaded = false;
+      },
+      this.handleLoadError,
+    );
   }
+  set intersecting(newValue: boolean) {
+    const old = this.#intersecting;
+    if (old !== newValue) {
+      this.#intersecting = newValue;
+      if (newValue) {
+        void this.store.loadBucket(this.bucketDate);
+      } else {
+        this.cancel();
+      }
+    }
+  }
+
+  get intersecting() {
+    return this.#intersecting;
+  }
+
+  get lastDateGroup() {
+    return this.dateGroups.at(-1);
+  }
+  getFirstAsset() {
+    return this.dateGroups[0]?.getFirstAsset();
+  }
+  getAssets() {
+    // eslint-disable-next-line unicorn/no-array-reduce
+    return this.dateGroups.reduce(
+      (accumulator: AssetResponseDto[], g: AssetDateGroup) => accumulator.concat(g.getAssets()),
+      [],
+    );
+  }
+
+  containsAssetId(id: string) {
+    for (const group of this.dateGroups) {
+      const index = group.intersetingAssets.findIndex((a) => a.id == id);
+      if (index !== -1) {
+        return true;
+      }
+    }
+    return false;
+  }
+
+  sortDateGroups() {
+    this.dateGroups.sort((a, b) => b.date.diff(a.date).milliseconds);
+  }
+
+  runAssetOperation(ids: Set<string>, operation: AssetOperation) {
+    if (ids.size === 0) {
+      return {
+        moveAssets: [] as MoveAsset[],
+        processedIds: new Set<string>(),
+        unprocessedIds: ids,
+        changedGeometry: false,
+      };
+    }
+    const { dateGroups } = this;
+    let combinedChangedGeometry = false;
+    let idsToProcess = new Set(ids);
+    const idsProcessed = new Set<string>();
+    const combinedMoveAssets: MoveAsset[][] = [];
+    let index = dateGroups.length;
+    while (index--) {
+      if (idsToProcess.size > 0) {
+        const group = dateGroups[index];
+        const { moveAssets, processedIds, changedGeometry } = group.runAssetOperation(ids, operation);
+        if (moveAssets.length > 0) {
+          combinedMoveAssets.push(moveAssets);
+        }
+        idsToProcess = idsToProcess.difference(processedIds);
+        for (const id of processedIds) {
+          idsProcessed.add(id);
+        }
+        combinedChangedGeometry = combinedChangedGeometry || changedGeometry;
+        if (group.intersetingAssets.length === 0) {
+          dateGroups.splice(index, 1);
+          combinedChangedGeometry = true;
+        }
+      }
+    }
+    return {
+      moveAssets: combinedMoveAssets.flat(),
+      unprocessedIds: idsToProcess,
+      processedIds: idsProcessed,
+      changedGeometry: combinedChangedGeometry,
+    };
+  }
+
+  // note - if the assets are not part of this bucket, they will not be added
+  addAssets(assets: AssetResponseDto[]) {
+    const lookupCache: {
+      [dayOfMonth: number]: AssetDateGroup;
+    } = {};
+    const unprocessedAssets: AssetResponseDto[] = [];
+    const changedDateGroups = new Set<AssetDateGroup>();
+    const newDateGroups = new Set<AssetDateGroup>();
+    for (const asset of assets) {
+      const date = DateTime.fromISO(asset.localDateTime).toUTC();
+      const month = date.get('month');
+      const year = date.get('year');
+      if (this.month === month && this.year === year) {
+        const day = date.get('day');
+        let dateGroup: AssetDateGroup | undefined = lookupCache[day];
+        if (!dateGroup) {
+          dateGroup = this.findDateGroupByDay(day);
+          if (dateGroup) {
+            lookupCache[day] = dateGroup;
+          }
+        }
+        if (dateGroup) {
+          const intersectingAsset = new IntersectingAsset(dateGroup, asset);
+          dateGroup.intersetingAssets.push(intersectingAsset);
+          changedDateGroups.add(dateGroup);
+        } else {
+          dateGroup = new AssetDateGroup(this, this.dateGroups.length, date, day);
+          dateGroup.intersetingAssets.push(new IntersectingAsset(dateGroup, asset));
+          this.dateGroups.push(dateGroup);
+          lookupCache[day] = dateGroup;
+          newDateGroups.add(dateGroup);
+        }
+      } else {
+        unprocessedAssets.push(asset);
+      }
+    }
+    for (const group of changedDateGroups) {
+      group.sortAssets();
+    }
+    for (const group of newDateGroups) {
+      group.sortAssets();
+    }
+    if (newDateGroups.size > 0) {
+      this.sortDateGroups();
+    }
+    return unprocessedAssets;
+  }
+  getRandomDateGroup() {
+    const random = Math.floor(Math.random() * this.dateGroups.length);
+    return this.dateGroups[random];
+  }
+
+  getRandomAsset() {
+    return this.getRandomDateGroup()?.getRandomAsset()?.asset;
+  }
+
   /** The svelte key for this view model object */
   get viewId() {
-    return this.store.viewId + '-' + this.bucketDate;
-  }
-  private init() {
-    // create a promise, and store its resolve/reject callbacks. The loadedSignal callback
-    // will be incoked when a bucket is loaded, fulfilling the promise. The canceledSignal
-    // callback will be called if the bucket is canceled before it was loaded, rejecting the
-    // promise.
-    this.complete = new Promise<void>((resolve, reject) => {
-      this.loadedSignal = resolve;
-      this.canceledSignal = reject;
-    }).catch(
-      () =>
-        // if no-one waits on complete, and its rejected a uncaught rejection message is logged.
-        // We this message with an empty reject handler, since waiting on a bucket is optional.
-        void 0,
-    );
-
-    this.measuredPromise = new Promise((resolve) => {
-      this.measuredSignal = resolve;
-    });
-
-    this.bucketDateFormattted = fromLocalDateTime(this.bucketDate)
-      .startOf('month')
-      .toJSDate()
-      .toLocaleString(get(locale), {
-        month: 'short',
-        year: 'numeric',
-        timeZone: 'UTC',
-      });
+    return this.bucketDate;
   }
 
-  private loadedSignal: (() => void) | undefined;
-  private canceledSignal: (() => void) | undefined;
-  measuredSignal: (() => void) | undefined;
+  set bucketHeight(height: number) {
+    const { store } = this;
+    const index = store.buckets.indexOf(this);
+    const bucketHeightDelta = height - this.#bucketHeight;
+    const prevBucket = store.buckets[index - 1];
+    if (prevBucket) {
+      this.#top = prevBucket.#top + prevBucket.#bucketHeight;
+    }
+    if (bucketHeightDelta) {
+      let cursor = index + 1;
+      while (cursor < store.buckets.length) {
+        const nextBucket = this.store.buckets[cursor];
+        nextBucket.#top += bucketHeightDelta;
+        cursor++;
+      }
+    }
+    this.#bucketHeight = height;
+    if (store.topIntersectingBucket) {
+      const currentIndex = store.buckets.indexOf(store.topIntersectingBucket);
+      // if the bucket is 'before' the last intersecting bucket in the sliding window
+      // then adjust the scroll position by the delta, to compensate for the bucket
+      // size adjustment
+      if (currentIndex > 0 && index <= currentIndex) {
+        store.compensateScrollCallback?.(bucketHeightDelta);
+      }
+    }
+  }
+  get bucketHeight() {
+    return this.#bucketHeight;
+  }
+
+  set top(top: number) {
+    this.#top = top;
+  }
+  get top() {
+    return this.#top + this.store.topSectionHeight;
+  }
+
+  handleLoadError(error: unknown) {
+    const _$t = get(t);
+    handleError(error, _$t('errors.failed_to_load_assets'));
+  }
+
+  findDateGroupByDay(dayOfMonth: number) {
+    return this.dateGroups.find((group) => group.dayOfMonth === dayOfMonth);
+  }
+
+  findAssetAbsolutePosition(assetId: string) {
+    for (const group of this.dateGroups) {
+      const intersectingAsset = group.intersetingAssets.find((asset) => asset.id === assetId);
+      if (intersectingAsset) {
+        return this.top + group.top + intersectingAsset.position!.top + HEADER;
+      }
+    }
+    return -1;
+  }
 
   cancel() {
-    if (this.isLoaded) {
-      return;
-    }
-    if (this.isPreventCancel) {
-      return;
-    }
-    this.cancelToken?.abort();
-    this.canceledSignal?.();
-    this.init();
-  }
-
-  loaded() {
-    this.loadedSignal?.();
-    this.isLoaded = true;
-  }
-
-  errored() {
-    this.canceledSignal?.();
-    this.init();
+    this.loader?.cancel();
   }
 }
 
 const isMismatched = (option: boolean | undefined, value: boolean): boolean =>
   option === undefined ? false : option !== value;
-
-const THUMBNAIL_HEIGHT = 235;
 
 interface AddAsset {
   type: 'add';
@@ -162,746 +504,719 @@ export const photoViewerImgElement = writable<HTMLImageElement | null>(null);
 
 type PendingChange = AddAsset | UpdateAsset | DeleteAsset | TrashAssets | UpdateStackAssets;
 
-export type BucketListener = (
-  event:
-    | ViewPortEvent
-    | BucketLoadEvent
-    | BucketLoadedEvent
-    | BucketCancelEvent
-    | BucketHeightEvent
-    | DateGroupIntersecting
-    | DateGroupHeightEvent,
-) => void;
-
-type ViewPortEvent = {
-  type: 'viewport';
-};
-type BucketLoadEvent = {
-  type: 'load';
-  bucket: AssetBucket;
-};
-type BucketLoadedEvent = {
-  type: 'loaded';
-  bucket: AssetBucket;
-};
-type BucketCancelEvent = {
-  type: 'cancel';
-  bucket: AssetBucket;
-};
-type BucketHeightEvent = {
-  type: 'bucket-height';
-  bucket: AssetBucket;
-  delta: number;
-};
-type DateGroupIntersecting = {
-  type: 'intersecting';
-  bucket: AssetBucket;
-  dateGroup: DateGroup;
-};
-type DateGroupHeightEvent = {
-  type: 'height';
-  bucket: AssetBucket;
-  dateGroup: DateGroup;
-  delta: number;
-  height: number;
+export type LiteBucket = {
+  bucketHeight: number;
+  assetCount: number;
+  bucketDate: string;
+  bucketDateFormattted: string;
 };
 
 export class AssetStore {
-  private assetToBucket: Record<string, AssetLookup> = $derived.by(() => {
-    const result: Record<string, AssetLookup> = {};
-    for (let index = 0; index < this.buckets.length; index++) {
-      const bucket = this.buckets[index];
-      for (let index_ = 0; index_ < bucket.assets.length; index_++) {
-        const asset = bucket.assets[index_];
-        result[asset.id] = { bucket, bucketIndex: index, assetIndex: index_ };
-      }
-    }
-    return result;
-  });
-  private pendingChanges: PendingChange[] = [];
-  private unsubscribers: Unsubscriber[] = [];
-  private options!: AssetApiGetTimeBucketsRequest;
-  viewport: Viewport = $state({
-    height: 0,
-    width: 0,
-  });
-  private initializedSignal!: () => void;
-  private store$ = writable(this);
-
-  /** The svelte key for this view model object */
-  viewId = generateId();
-
-  lastScrollTime: number = $state(0);
-
-  // subscribe = this.store$.subscribe;
-  /**
-   * A promise that resolves once the store is initialized.
-   */
-  private complete!: Promise<void>;
-  taskManager = new AssetGridTaskManager(this);
-  initialized = $state(false);
-  timelineHeight = $state(0);
+  // --- public ----
+  isInitialized = $state(false);
   buckets: AssetBucket[] = $state([]);
-  assets: AssetResponseDto[] = $derived.by(() => {
-    return this.buckets.flatMap(({ assets }) => assets);
-  });
+  topSectionHeight = $state(0);
+  timelineHeight = $derived(
+    this.buckets.reduce((accumulator, b) => accumulator + b.bucketHeight, 0) + this.topSectionHeight,
+  );
+
+  // todo - name this better
   albumAssets: Set<string> = new SvelteSet();
-  pendingScrollBucket: AssetBucket | undefined = $state();
-  pendingScrollAssetId: string | undefined = $state();
-  maxBucketAssets = $state(0);
 
-  private listeners: BucketListener[] = [];
+  // -- for scrubber only
+  scrubberBuckets: LiteBucket[] = $state([]);
+  scrubberTimelineHeight: number = $state(0);
 
-  constructor(
-    options: AssetStoreOptions,
-    private albumId?: string,
-  ) {
-    this.setOptions(options);
-    this.createInitializationSignal();
-    this.store$.set(this);
+  // -- should be private, but used by AssetBucket
+  compensateScrollCallback: ((delta: number) => void) | undefined;
+  topIntersectingBucket: AssetBucket | undefined = $state();
+
+  visibleWindow = $derived.by(() => ({
+    top: this.#scrollTop,
+    bottom: this.#scrollTop + this.viewportHeight,
+  }));
+
+  initTask = new CancellableTask(
+    () => {
+      this.isInitialized = true;
+      this.connect();
+    },
+    () => {
+      this.disconnect();
+      this.isInitialized = false;
+    },
+    () => void 0,
+  );
+
+  // --- private
+  static #INIT_OPTIONS = {};
+  #viewportHeight = $state(0);
+  #viewportWidth = $state(0);
+  #scrollTop = $state(0);
+  #pendingChanges: PendingChange[] = [];
+  #unsubscribers: Unsubscriber[] = [];
+
+  #options: AssetStoreOptions = AssetStore.#INIT_OPTIONS;
+
+  #scrolling = $state(false);
+  #suspendTransitions = $state(false);
+  #resetScrolling = debounce(() => (this.#scrolling = false), 1000);
+  #resetSuspendTransitions = debounce(() => (this.suspendTransitions = false), 1000);
+
+  constructor() {}
+
+  set scrolling(value: boolean) {
+    this.#scrolling = value;
+    if (value) {
+      this.suspendTransitions = true;
+      this.#resetScrolling();
+    }
   }
 
-  private setOptions(options: AssetStoreOptions) {
-    this.options = { ...options, size: TimeBucketSize.Month };
+  get scrolling() {
+    return this.#scrolling;
   }
 
-  private createInitializationSignal() {
-    // create a promise, and store its resolve callbacks. The initializedSignal callback
-    // will be invoked when a the assetstore is initialized.
-    this.complete = new Promise<void>((resolve) => {
-      this.initializedSignal = resolve;
-    }).catch(() => void 0);
+  set suspendTransitions(value: boolean) {
+    this.#suspendTransitions = value;
+    if (value) {
+      this.#resetSuspendTransitions();
+    }
   }
 
-  private addPendingChanges(...changes: PendingChange[]) {
-    // prevent websocket events from happening before local client events
-    setTimeout(() => {
-      this.pendingChanges.push(...changes);
-      this.processPendingChanges();
-    }, 1000);
+  get suspendTransitions() {
+    return this.#suspendTransitions;
+  }
+
+  set viewportWidth(value: number) {
+    const changed = value !== this.#viewportWidth;
+    this.#viewportWidth = value;
+    this.suspendTransitions = true;
+    // side-effect - its ok!
+    void this.#updateViewportGeometry(changed);
+  }
+
+  get viewportWidth() {
+    return this.#viewportWidth;
+  }
+
+  set viewportHeight(value: number) {
+    this.#viewportHeight = value;
+    this.#suspendTransitions = true;
+    // side-effect - its ok!
+    void this.#updateViewportGeometry(false);
+  }
+
+  get viewportHeight() {
+    return this.#viewportHeight;
+  }
+
+  getAssets() {
+    return this.buckets.flatMap((bucket) => bucket.getAssets());
+  }
+
+  #addPendingChanges(...changes: PendingChange[]) {
+    this.#pendingChanges.push(...changes);
+    this.#processPendingChanges();
   }
 
   connect() {
-    this.unsubscribers.push(
-      websocketEvents.on('on_upload_success', (_) => {
-        // TODO!: Temporarily disable to avoid flashing effect of the timeline
-        // this.addPendingChanges({ type: 'add', values: [asset] });
-      }),
-      websocketEvents.on('on_asset_trash', (ids) => {
-        this.addPendingChanges({ type: 'trash', values: ids });
-      }),
-      websocketEvents.on('on_asset_update', (asset) => {
-        this.addPendingChanges({ type: 'update', values: [asset] });
-      }),
-      websocketEvents.on('on_asset_delete', (id: string) => {
-        this.addPendingChanges({ type: 'delete', values: [id] });
-      }),
+    this.#unsubscribers.push(
+      websocketEvents.on('on_upload_success', (asset) => this.#addPendingChanges({ type: 'add', values: [asset] })),
+      websocketEvents.on('on_asset_trash', (ids) => this.#addPendingChanges({ type: 'trash', values: ids })),
+      websocketEvents.on('on_asset_update', (asset) => this.#addPendingChanges({ type: 'update', values: [asset] })),
+      websocketEvents.on('on_asset_delete', (id: string) => this.#addPendingChanges({ type: 'delete', values: [id] })),
     );
   }
 
   disconnect() {
-    for (const unsubscribe of this.unsubscribers) {
+    for (const unsubscribe of this.#unsubscribers) {
       unsubscribe();
     }
-    this.unsubscribers = [];
+    this.#unsubscribers = [];
   }
 
-  private getPendingChangeBatches() {
-    const batches: PendingChange[] = [];
-    let batch: PendingChange | undefined;
-
-    for (const { type, values: _values } of this.pendingChanges) {
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      const values = _values as any[];
-
-      if (batch && batch.type !== type) {
-        batches.push(batch);
-        batch = undefined;
-      }
-
-      if (batch) {
-        batch.values.push(...values);
-      } else {
-        batch = { type, values };
-      }
-    }
-
-    if (batch) {
-      batches.push(batch);
-    }
-
-    return batches;
-  }
-
-  processPendingChanges = throttle(() => {
-    for (const { type, values } of this.getPendingChangeBatches()) {
+  #getPendingChangeBatches() {
+    const batch: {
+      add: AssetResponseDto[];
+      update: AssetResponseDto[];
+      remove: string[];
+    } = {
+      add: [],
+      update: [],
+      remove: [],
+    };
+    for (const { type, values } of this.#pendingChanges) {
       switch (type) {
         case 'add': {
-          this.addAssets(values);
+          batch.add.push(...values);
+
           break;
         }
-
         case 'update': {
-          this.updateAssets(values);
+          batch.update.push(...values);
+
           break;
         }
-
+        case 'delete':
         case 'trash': {
-          if (!this.options.isTrashed) {
-            this.removeAssets(values);
-          }
-          break;
-        }
+          batch.remove.push(...values);
 
-        case 'delete': {
-          this.removeAssets(values);
           break;
         }
+        // No default
       }
     }
+    return batch;
+  }
 
-    this.pendingChanges = [];
-    // this.emit(true);
+  // todo: this should probably be a method isteat
+  #findBucketForAsset(id: string) {
+    for (const bucket of this.buckets) {
+      if (bucket.containsAssetId(id)) {
+        return bucket;
+      }
+    }
+  }
+
+  updateSlidingWindow(scrollTop: number) {
+    this.#scrollTop = scrollTop;
+    this.updateIntersections();
+  }
+
+  updateIntersections() {
+    if (!this.isInitialized || this.visibleWindow.bottom === this.visibleWindow.top) {
+      return;
+    }
+    let topIntersectingBucket = undefined;
+    for (const bucket of this.buckets) {
+      this.#updateIntersection(bucket);
+      if (!topIntersectingBucket && bucket.intersecting) {
+        topIntersectingBucket = bucket;
+      }
+    }
+    if (this.topIntersectingBucket !== topIntersectingBucket) {
+      this.topIntersectingBucket = topIntersectingBucket;
+    }
+  }
+
+  #updateIntersection(bucket: AssetBucket) {
+    const bucketTop = bucket.top;
+    const bucketBottom = bucketTop + bucket.bucketHeight;
+    const topWindow = this.visibleWindow.top - INTERSECTION_EXPAND_TOP;
+    const bottomWindow = this.visibleWindow.bottom + INTERSECTION_EXPAND_BOTTOM;
+
+    // a bucket intersections if
+    // 1) bucket's bottom is in the visible range -or-
+    // 2) bucket's bottom is in the visible range -or-
+    // 3) bucket's top is above visible range and bottom is below visible range
+    bucket.intersecting =
+      (bucketTop >= topWindow && bucketTop < bottomWindow) ||
+      (bucketBottom >= topWindow && bucketBottom < bottomWindow) ||
+      (bucketTop < topWindow && bucketBottom >= bottomWindow);
+  }
+
+  #processPendingChanges = throttle(() => {
+    const { add, update, remove } = this.#getPendingChangeBatches();
+    if (add.length > 0) {
+      this.addAssets(add);
+    }
+    if (update.length > 0) {
+      this.updateAssets(update);
+    }
+    if (remove.length > 0) {
+      this.removeAssets(remove);
+    }
+    this.#pendingChanges = [];
   }, 2500);
 
-  addListener(bucketListener: BucketListener) {
-    this.listeners.push(bucketListener);
-  }
-  removeListener(bucketListener: BucketListener) {
-    this.listeners = this.listeners.filter((l) => l != bucketListener);
-  }
-  private notifyListeners(
-    event:
-      | ViewPortEvent
-      | BucketLoadEvent
-      | BucketLoadedEvent
-      | BucketCancelEvent
-      | BucketHeightEvent
-      | DateGroupIntersecting
-      | DateGroupHeightEvent,
-  ) {
-    for (const fn of this.listeners) {
-      fn(event);
-    }
-  }
-  async init({ bucketListener }: { bucketListener?: BucketListener } = {}) {
-    if (this.initialized) {
-      throw 'Can only init once';
-    }
-    if (!getJustifiedLayoutFromAssets) {
-      const module = await import('$lib/utils/layout-utils');
-      getJustifiedLayoutFromAssets = module.getJustifiedLayoutFromAssets;
-    }
-
-    if (bucketListener) {
-      this.addListener(bucketListener);
-    }
-    await this.initialiazeTimeBuckets();
+  setCompensateScrollCallback(compensateScrollCallback?: (delta: number) => void) {
+    this.compensateScrollCallback = compensateScrollCallback;
   }
 
-  async initialiazeTimeBuckets() {
-    this.timelineHeight = 0;
-    this.buckets = [];
-    this.albumAssets.clear();
-
+  async #initialiazeTimeBuckets() {
     const timebuckets = await getTimeBuckets({
-      ...this.options,
+      ...this.#options,
+      size: TimeBucketSize.Month,
       key: getKey(),
     });
-    this.buckets = timebuckets.map(
-      (bucket) => new AssetBucket({ store: this, bucketDate: bucket.timeBucket, initialCount: bucket.count }),
-    );
 
-    this.initializedSignal();
-    this.initialized = true;
+    this.buckets = timebuckets.map((bucket) => {
+      const utcDate = DateTime.fromISO(bucket.timeBucket).toUTC();
+      return new AssetBucket(this, utcDate, bucket.count);
+    });
+    this.albumAssets.clear();
+    this.#updateViewportGeometry(false);
   }
 
+  /**
+   * If the timeline query options change (i.e. albumId, isArchived, isFavorite, etc)
+   * call this method to recreate all buckets based on the new options.
+   *
+   * @param options The query options for time bucket queries.
+   */
   async updateOptions(options: AssetStoreOptions) {
-    // Make sure to re-initialize if the personId changes
-    const needsReinitializing = this.options.personId !== options.personId;
-    if (!this.initialized && !needsReinitializing) {
-      this.setOptions(options);
+    if (options.deferInit) {
       return;
     }
-
-    // Make sure to re-initialize if the tagId changes
-    if (this.options.tagId === options.tagId) {
-      this.setOptions(options);
+    if (this.#options !== AssetStore.#INIT_OPTIONS && isEqual(this.#options, options)) {
       return;
     }
-    // TODO: don't call updateObjects frequently after
-    // init - cancelation of the initialize tasks isn't
-    // performed right now, and will cause issues if
-    // multiple updateOptions() calls are interleved.
-    await this.complete;
-    this.taskManager.destroy();
-    this.taskManager = new AssetGridTaskManager(this);
-    this.initialized = false;
-    this.viewId = generateId();
-    this.createInitializationSignal();
-    this.setOptions(options);
-    await this.initialiazeTimeBuckets();
-    // this.emit(true);
-    await this.initialLayout(true);
+    await this.initTask.reset();
+    await this.#init(options);
+    this.#updateViewportGeometry(false);
   }
 
+  async #init(options: AssetStoreOptions) {
+    // doing the following outside of the task reduces flickr
+    this.isInitialized = false;
+    this.buckets = [];
+    this.albumAssets.clear();
+    await this.initTask.execute(async () => {
+      this.#options = options;
+      await this.#initialiazeTimeBuckets();
+    }, true);
+  }
   public destroy() {
-    this.taskManager.destroy();
-    this.listeners = [];
-    this.initialized = false;
+    this.disconnect();
+    this.isInitialized = false;
   }
 
-  async updateViewport(viewport: Viewport, force?: boolean) {
+  async updateViewport(viewport: Viewport) {
     if (viewport.height === 0 && viewport.width === 0) {
       return;
     }
-    if (!force && this.viewport.height === viewport.height && this.viewport.width === viewport.width) {
+
+    if (this.viewportHeight === viewport.height && this.viewportWidth === viewport.width) {
       return;
     }
-    await this.complete;
-    // changing width invalidates the actual height, and needs to be remeasured, since width changes causes
-    // layout reflows.
-    const changedWidth = this.viewport.width != viewport.width;
-    this.viewport = { ...viewport };
-    await this.initialLayout(changedWidth);
-  }
 
-  private async initialLayout(changedWidth: boolean) {
-    for (const bucket of this.buckets) {
-      this.updateGeometry(bucket, changedWidth);
-    }
-    this.timelineHeight = this.buckets.reduce((accumulator, b) => accumulator + b.bucketHeight, 0);
-
-    const loaders = [];
-    let height = 0;
-    for (const bucket of this.buckets) {
-      if (height >= this.viewport.height) {
-        break;
-      }
-      height += bucket.bucketHeight;
-      loaders.push(this.loadBucket(bucket.bucketDate));
-    }
-    await Promise.all(loaders);
-    this.notifyListeners({ type: 'viewport' });
-  }
-
-  private updateGeometry(bucket: AssetBucket, invalidateHeight: boolean) {
-    if (invalidateHeight) {
-      bucket.isBucketHeightActual = false;
-      bucket.measured = false;
-      for (const assetGroup of bucket.dateGroups) {
-        assetGroup.heightActual = false;
+    // special case updateViewport before or soon after call to updateOptions
+    if (!this.initTask.executed) {
+      // eslint-disable-next-line unicorn/prefer-ternary
+      if (this.initTask.loading) {
+        await this.initTask.waitUntilCompletion();
+      } else {
+        // not executed and not loaded means we should init now, and init will
+        // also update geometry so just return after
+        await this.#init(this.#options);
       }
     }
-    const viewportWidth = this.viewport.width;
-    if (!bucket.isBucketHeightActual) {
-      const unwrappedWidth = (3 / 2) * bucket.bucketCount * THUMBNAIL_HEIGHT * (7 / 10);
-      const rows = Math.ceil(unwrappedWidth / viewportWidth);
-      const height = 51 + Math.max(1, rows) * THUMBNAIL_HEIGHT;
 
-      this.setBucketHeight(bucket, height, false);
+    // changing width affects the actual height, and needs to re-layout
+    const changedWidth = viewport.width !== this.viewportWidth;
+    this.viewportHeight = viewport.height;
+    this.viewportWidth = viewport.width;
+    this.#updateViewportGeometry(changedWidth);
+  }
+
+  #updateViewportGeometry(changedWidth: boolean) {
+    if (!this.isInitialized) {
+      return;
     }
-    const layoutOptions = {
+    if (this.viewportWidth === 0 || this.viewportHeight === 0) {
+      return;
+    }
+    for (const bucket of this.buckets) {
+      this.#updateGeometry(bucket, changedWidth);
+    }
+    this.updateIntersections();
+    this.#createScrubBuckets();
+  }
+
+  #createScrubBuckets() {
+    this.scrubberBuckets = this.buckets.map((bucket) => ({
+      assetCount: bucket.bucketCount,
+      bucketDate: bucket.bucketDate,
+      bucketDateFormattted: bucket.bucketDateFormatted,
+      bucketHeight: bucket.bucketHeight,
+    }));
+    this.scrubberTimelineHeight = this.timelineHeight;
+  }
+
+  createLayoutOptions() {
+    const viewportWidth = this.viewportWidth;
+    return {
       spacing: 2,
       heightTolerance: 0.15,
       rowHeight: 235,
       rowWidth: Math.floor(viewportWidth),
     };
-    for (const assetGroup of bucket.dateGroups) {
-      if (!assetGroup.heightActual) {
-        const unwrappedWidth = (3 / 2) * assetGroup.assets.length * THUMBNAIL_HEIGHT * (7 / 10);
-        const rows = Math.ceil(unwrappedWidth / this.viewport.width);
-        const height = rows * THUMBNAIL_HEIGHT;
-        assetGroup.height = height;
-      }
-
-      assetGroup.geometry = getJustifiedLayoutFromAssets(assetGroup.assets, layoutOptions);
+  }
+  #updateGeometry(bucket: AssetBucket, invalidateHeight: boolean) {
+    if (invalidateHeight) {
+      bucket.isBucketHeightActual = false;
     }
+    if (!bucket.isLoaded) {
+      // optimize - if bucket already has data, no need to create estimates
+      const viewportWidth = this.viewportWidth;
+      if (!bucket.isBucketHeightActual) {
+        const unwrappedWidth = (3 / 2) * bucket.bucketCount * THUMBNAIL_HEIGHT * (7 / 10);
+        const rows = Math.ceil(unwrappedWidth / viewportWidth);
+        const height = 51 + Math.max(1, rows) * THUMBNAIL_HEIGHT;
+        bucket.bucketHeight = height;
+      }
+      return;
+    }
+    this.#layoutBucket(bucket);
   }
 
-  async loadBucket(bucketDate: string, options: { preventCancel?: boolean; pending?: boolean } = {}): Promise<void> {
-    const bucket = this.getBucketByDate(bucketDate);
+  #layoutBucket(bucket: AssetBucket) {
+    // these are top offsets, for each row
+    let cummulativeHeight = 0;
+    // these are left offsets of each group, for each row
+    let cummulativeWidth = 0;
+    let lastRowHeight = 0;
+    let lastRow = 0;
+
+    let dateGroupRow = 0;
+    let dateGroupCol = 0;
+
+    const rowSpaceRemaining: number[] = Array.from({ length: bucket.dateGroups.length });
+    rowSpaceRemaining.fill(this.viewportWidth, 0, bucket.dateGroups.length);
+    const options = this.createLayoutOptions();
+    for (const assetGroup of bucket.dateGroups) {
+      assetGroup.layout(options);
+      rowSpaceRemaining[dateGroupRow] -= assetGroup.width - 1;
+      if (dateGroupCol > 0) {
+        rowSpaceRemaining[dateGroupRow] -= GAP;
+      }
+      if (rowSpaceRemaining[dateGroupRow] >= 0) {
+        assetGroup.row = dateGroupRow;
+        assetGroup.col = dateGroupCol;
+        assetGroup.left = cummulativeWidth;
+        assetGroup.top = cummulativeHeight;
+
+        dateGroupCol++;
+
+        cummulativeWidth += assetGroup.width + GAP;
+      } else {
+        // starting a new row, we need to update the last col of the previous row
+        cummulativeWidth = 0;
+        dateGroupRow++;
+        dateGroupCol = 0;
+        assetGroup.row = dateGroupRow;
+        assetGroup.col = dateGroupCol;
+        assetGroup.left = cummulativeWidth;
+
+        rowSpaceRemaining[dateGroupRow] -= assetGroup.width;
+        dateGroupCol++;
+        cummulativeHeight += lastRowHeight;
+        assetGroup.top = cummulativeHeight;
+        cummulativeWidth += assetGroup.width + GAP;
+        lastRow = assetGroup.row - 1;
+      }
+      lastRowHeight = assetGroup.height + HEADER;
+    }
+    if (lastRow === 0 || lastRow !== bucket.lastDateGroup?.row) {
+      cummulativeHeight += lastRowHeight;
+    }
+
+    bucket.bucketHeight = cummulativeHeight;
+    bucket.isBucketHeightActual = true;
+  }
+
+  async loadBucket(bucketDate: string, options?: { cancelable: boolean }): Promise<void> {
+    let cancelable = true;
+    if (options) {
+      cancelable = options.cancelable;
+    }
+
+    const date = DateTime.fromISO(bucketDate).toUTC();
+    const year = date.get('year');
+    const month = date.get('month');
+    const bucket = this.getBucketByDate(year, month);
     if (!bucket) {
       return;
     }
-    if (bucket.isLoaded) {
-      // already loaded
+
+    if (bucket.loader?.executed) {
       return;
     }
 
-    if (bucket.cancelToken != null && bucket.bucketCount !== bucket.assets.length) {
-      // if promise is pending, and preventCancel is requested, then don't overwrite it
-      if (!bucket.isPreventCancel && options.preventCancel) {
-        bucket.isPreventCancel = options.preventCancel;
-      }
-      await bucket.complete;
-      return;
-    }
-
-    if (options.pending) {
-      this.pendingScrollBucket = bucket;
-    }
-    this.notifyListeners({ type: 'load', bucket });
-    bucket.isPreventCancel = !!options.preventCancel;
-    const cancelToken = (bucket.cancelToken = new AbortController());
-    try {
+    const result = await bucket.loader?.execute(async (signal: AbortSignal) => {
       const assets = await getTimeBucket(
         {
-          ...this.options,
+          ...this.#options,
           timeBucket: bucketDate,
+          size: TimeBucketSize.Month,
           key: getKey(),
         },
-        { signal: cancelToken.signal },
+        { signal },
       );
-
-      if (cancelToken.signal.aborted) {
-        this.notifyListeners({ type: 'cancel', bucket });
-        return;
-      }
-
-      if (this.albumId) {
-        const albumAssets = await getTimeBucket(
-          {
-            albumId: this.albumId,
-            timeBucket: bucketDate,
-            size: this.options.size,
-            key: getKey(),
-          },
-          { signal: cancelToken.signal },
-        );
-        if (cancelToken.signal.aborted) {
-          this.notifyListeners({ type: 'cancel', bucket });
-          return;
+      if (assets) {
+        if (this.#options.timelineAlbumId) {
+          const albumAssets = await getTimeBucket(
+            {
+              albumId: this.#options.timelineAlbumId,
+              timeBucket: bucketDate,
+              size: TimeBucketSize.Month,
+              key: getKey(),
+            },
+            { signal },
+          );
+          for (const asset of albumAssets) {
+            this.albumAssets.add(asset.id);
+          }
         }
-        for (const asset of albumAssets) {
-          this.albumAssets.add(asset.id);
+
+        const unprocessed = bucket.addAssets(assets);
+        if (unprocessed.length > 0) {
+          console.error(
+            `Warning: getTimeBucket API returning assets not in requested month: ${bucket.bucketDate}, ${JSON.stringify(unprocessed.map((a) => ({ id: a.id, localDateTime: a.localDateTime })))}`,
+          );
         }
+        this.#layoutBucket(bucket);
       }
-
-      bucket.assets = assets;
-      bucket.dateGroups = splitBucketIntoDateGroups(bucket, get(locale));
-      this.updateGeometry(bucket, true);
-      this.timelineHeight = this.buckets.reduce((accumulator, b) => accumulator + b.bucketHeight, 0);
-      bucket.loaded();
-      this.notifyListeners({ type: 'loaded', bucket });
-    } catch (error) {
-      /* eslint-disable-next-line  @typescript-eslint/no-explicit-any */
-      if ((error as any).name === 'AbortError') {
-        return;
-      }
-      const _$t = get(t);
-      handleError(error, _$t('errors.failed_to_load_assets'));
-      bucket.errored();
-    } finally {
-      bucket.cancelToken = undefined;
+    }, cancelable);
+    if (result === 'LOADED') {
+      this.#updateIntersection(bucket);
     }
-  }
-
-  setBucketHeight(bucket: AssetBucket, newHeight: number, isActualHeight: boolean) {
-    const delta = newHeight - bucket.bucketHeight;
-    bucket.isBucketHeightActual = isActualHeight;
-    bucket.bucketHeight = newHeight;
-    this.timelineHeight += delta;
-    this.notifyListeners({ type: 'bucket-height', bucket, delta });
-  }
-
-  updateBucket(bucketDate: string, properties: { height?: number; intersecting?: boolean; measured?: boolean }) {
-    const bucket = this.getBucketByDate(bucketDate);
-    if (!bucket) {
-      return {};
-    }
-    const delta = 0;
-    if ('height' in properties) {
-      this.setBucketHeight(bucket, properties.height!, true);
-    }
-    if ('intersecting' in properties) {
-      bucket.intersecting = properties.intersecting!;
-    }
-    if ('measured' in properties) {
-      if (properties.measured) {
-        bucket.measuredSignal?.();
-      }
-      bucket.measured = properties.measured!;
-    }
-    return { delta };
-  }
-
-  updateBucketDateGroup(
-    bucket: AssetBucket,
-    dateGroup: DateGroup,
-    properties: { height?: number; intersecting?: boolean },
-  ) {
-    let delta = 0;
-    if ('height' in properties) {
-      const height = properties.height!;
-      if (height > 0) {
-        delta = height - dateGroup.height;
-        dateGroup.heightActual = true;
-        dateGroup.height = height;
-        this.notifyListeners({ type: 'height', bucket, dateGroup, delta, height });
-      }
-    }
-    if ('intersecting' in properties) {
-      dateGroup.intersecting = properties.intersecting!;
-      if (dateGroup.intersecting) {
-        this.notifyListeners({ type: 'intersecting', bucket, dateGroup });
-      }
-    }
-    return { delta };
   }
 
   addAssets(assets: AssetResponseDto[]) {
     const assetsToUpdate: AssetResponseDto[] = [];
-    const assetsToAdd: AssetResponseDto[] = [];
 
     for (const asset of assets) {
-      if (
-        this.assetToBucket[asset.id] ||
-        this.options.userId ||
-        this.options.personId ||
-        this.options.albumId ||
-        this.isExcluded(asset)
-      ) {
-        // If asset is already in the bucket we don't need to recalculate
-        // asset store containers
-        assetsToUpdate.push(asset);
-      } else {
-        assetsToAdd.push(asset);
+      if (this.isExcluded(asset)) {
+        continue;
       }
+      assetsToUpdate.push(asset);
     }
 
-    this.updateAssets(assetsToUpdate);
-    this.addAssetsToBuckets(assetsToAdd);
+    const notUpdated = this.updateAssets(assetsToUpdate);
+    this.#addAssetsToBuckets([...notUpdated]);
   }
 
-  private addAssetsToBuckets(assets: AssetResponseDto[]) {
+  #addAssetsToBuckets(assets: AssetResponseDto[]) {
     if (assets.length === 0) {
       return;
     }
     const updatedBuckets = new Set<AssetBucket>();
+    const updatedDateGroups = new Set<AssetDateGroup>();
 
     for (const asset of assets) {
-      const timeBucket = DateTime.fromISO(asset.localDateTime).toUTC().startOf('month').toString();
-      let bucket = this.getBucketByDate(timeBucket);
+      const utc = DateTime.fromISO(asset.localDateTime).toUTC().startOf('month');
+      const year = utc.get('year');
+      const month = utc.get('month');
+      let bucket = this.getBucketByDate(year, month);
 
       if (!bucket) {
-        bucket = new AssetBucket({ store: this, bucketDate: timeBucket, bucketHeight: THUMBNAIL_HEIGHT });
+        bucket = new AssetBucket(this, utc, 1);
         this.buckets.push(bucket);
       }
-
-      bucket.assets.push(asset);
+      bucket.addAssets([asset]);
       updatedBuckets.add(bucket);
     }
 
-    this.buckets = this.buckets.sort((a, b) => {
-      const aDate = DateTime.fromISO(a.bucketDate).toUTC();
-      const bDate = DateTime.fromISO(b.bucketDate).toUTC();
-      return bDate.diff(aDate).milliseconds;
+    this.buckets.sort((a, b) => {
+      return a.year === b.year ? b.month - a.month : b.year - a.year;
     });
 
-    for (const bucket of updatedBuckets) {
-      bucket.assets.sort((a, b) => {
-        const aDate = DateTime.fromISO(a.fileCreatedAt).toUTC();
-        const bDate = DateTime.fromISO(b.fileCreatedAt).toUTC();
-        return bDate.diff(aDate).milliseconds;
-      });
-      bucket.dateGroups = splitBucketIntoDateGroups(bucket, get(locale));
-      this.updateGeometry(bucket, true);
+    for (const dateGroup of updatedDateGroups) {
+      dateGroup.sortAssets();
     }
+    for (const bucket of updatedBuckets) {
+      bucket.sortDateGroups();
+      this.#updateGeometry(bucket, true);
+    }
+    this.updateIntersections();
   }
 
-  getBucketByDate(bucketDate: string): AssetBucket | null {
-    return this.buckets.find((bucket) => bucket.bucketDate === bucketDate) || null;
+  getBucketByDate(year: number, month: number): AssetBucket | undefined {
+    return this.buckets.find((bucket) => bucket.year === year && bucket.month === month);
   }
 
-  async findAndLoadBucketAsPending(id: string) {
-    const bucketInfo = this.assetToBucket[id];
-    let bucket: AssetBucket | null = bucketInfo?.bucket ?? null;
+  async findBucketForAsset(id: string) {
+    await this.initTask.waitUntilCompletion();
+    let bucket = this.#findBucketForAsset(id);
     if (!bucket) {
       const asset = await getAssetInfo({ id });
       if (!asset || this.isExcluded(asset)) {
         return;
       }
-      bucket = await this.loadBucketAtTime(asset.localDateTime, { preventCancel: true, pending: true });
+      bucket = await this.#loadBucketAtTime(asset.localDateTime, { cancelable: false });
     }
 
-    if (bucket && bucket.assets.some((a) => a.id === id)) {
-      this.pendingScrollBucket = bucket;
-      this.pendingScrollAssetId = id;
+    if (bucket && bucket?.containsAssetId(id)) {
       return bucket;
     }
   }
 
-  /* Must be paired with matching clearPendingScroll() call */
-  async scheduleScrollToAssetId(scrollTarget: AssetGridRouteSearchParams, onFailure: () => void) {
-    try {
-      const { at: assetId } = scrollTarget;
-      if (assetId) {
-        await this.complete;
-        const bucket = await this.findAndLoadBucketAsPending(assetId);
-        if (bucket) {
-          return;
-        }
-      }
-    } catch {
-      // failure
-    }
-    onFailure();
-  }
-
-  clearPendingScroll() {
-    this.pendingScrollBucket = undefined;
-    this.pendingScrollAssetId = undefined;
-  }
-
-  private async loadBucketAtTime(localDateTime: string, options: { preventCancel?: boolean; pending?: boolean }) {
+  async #loadBucketAtTime(localDateTime: string, options?: { cancelable: boolean }) {
     let date = fromLocalDateTime(localDateTime);
-    if (this.options.size == TimeBucketSize.Month) {
-      date = date.set({ day: 1, hour: 0, minute: 0, second: 0, millisecond: 0 });
-    } else if (this.options.size == TimeBucketSize.Day) {
-      date = date.set({ hour: 0, minute: 0, second: 0, millisecond: 0 });
-    }
+    // Only support TimeBucketSize.Month
+    date = date.set({ day: 1, hour: 0, minute: 0, second: 0, millisecond: 0 });
     const iso = date.toISO()!;
+    const year = date.get('year');
+    const month = date.get('month');
     await this.loadBucket(iso, options);
-    return this.getBucketByDate(iso);
+    return this.getBucketByDate(year, month);
   }
 
-  private async getBucketInfoForAsset(
-    { id, localDateTime }: Pick<AssetResponseDto, 'id' | 'localDateTime'>,
-    options: { preventCancel?: boolean; pending?: boolean } = {},
-  ) {
-    const bucketInfo = this.assetToBucket[id];
+  async #getBucketInfoForAsset(asset: AssetResponseDto, options?: { cancelable: boolean }) {
+    const bucketInfo = this.#findBucketForAsset(asset.id);
     if (bucketInfo) {
       return bucketInfo;
     }
-    await this.loadBucketAtTime(localDateTime, options);
-    return this.assetToBucket[id] || null;
+    await this.#loadBucketAtTime(asset.localDateTime, options);
+    return this.#findBucketForAsset(asset.id);
   }
 
   getBucketIndexByAssetId(assetId: string) {
-    return this.assetToBucket[assetId]?.bucketIndex ?? null;
+    return this.#findBucketForAsset(assetId);
   }
 
-  async getRandomAsset(): Promise<AssetResponseDto | null> {
-    let index = Math.floor(
-      Math.random() * this.buckets.reduce((accumulator, bucket) => accumulator + bucket.bucketCount, 0),
-    );
-    for (const bucket of this.buckets) {
-      if (index < bucket.bucketCount) {
-        await this.loadBucket(bucket.bucketDate);
-        return bucket.assets[index] || null;
-      }
+  async getRandomBucket() {
+    const random = Math.floor(Math.random() * this.buckets.length);
+    const bucket = this.buckets[random];
+    await this.loadBucket(bucket.bucketDate, { cancelable: false });
+    return bucket;
+  }
 
-      index -= bucket.bucketCount;
+  async getRandomAsset() {
+    const bucket = await this.getRandomBucket();
+    return bucket?.getRandomAsset();
+  }
+
+  // runs op on assets, returns unprocessed
+  #runAssetOperation(ids: Set<string>, operation: AssetOperation) {
+    if (ids.size === 0) {
+      return { processedIds: new Set(), unprocessedIds: ids, changedGeometry: false };
     }
 
-    return null;
+    const changedBuckets = new Set<AssetBucket>();
+    let idsToProcess = new Set(ids);
+    const idsProcessed = new Set<string>();
+    const combinedMoveAssets: { asset: AssetResponseDto; year: number; month: number }[][] = [];
+    for (const bucket of this.buckets) {
+      if (idsToProcess.size > 0) {
+        const { moveAssets, processedIds, changedGeometry } = bucket.runAssetOperation(idsToProcess, operation);
+        if (moveAssets.length > 0) {
+          combinedMoveAssets.push(moveAssets);
+        }
+        idsToProcess = idsToProcess.difference(processedIds);
+        for (const id of processedIds) {
+          idsProcessed.add(id);
+        }
+        if (changedGeometry) {
+          changedBuckets.add(bucket);
+          break;
+        }
+      }
+    }
+    if (combinedMoveAssets.length > 0) {
+      this.#addAssetsToBuckets(combinedMoveAssets.flat().map((a) => a.asset));
+    }
+    const changedGeometry = changedBuckets.size > 0;
+    for (const bucket of changedBuckets) {
+      this.#updateGeometry(bucket, true);
+    }
+    if (changedGeometry) {
+      this.updateIntersections();
+    }
+    return { unprocessedIds: idsToProcess, processedIds: idsProcessed, changedGeometry };
+  }
+
+  /**
+   * Runs a callback on a list of asset ids. The assets in the AssetStore are reactive -
+   * any change to the asset (i.e. changing isFavorite, isArchived, etc) will automatically
+   * cause the UI to update with no further actions needed. Changing the date of an asset
+   * will automatically move it to another bucket if needed. Removing the asset will remove
+   * it from any view that is showing it.
+   *
+   * @param ids to run the operation on
+   * @param operation callback to update the specified asset ids
+   */
+  updateAssetOperation(ids: string[], operation: AssetOperation) {
+    this.#runAssetOperation(new Set(ids), operation);
   }
 
   updateAssets(assets: AssetResponseDto[]) {
-    if (assets.length === 0) {
-      return;
-    }
-    const assetsToRecalculate: AssetResponseDto[] = [];
-
-    for (const _asset of assets) {
-      const asset = this.assets.find((asset) => asset.id === _asset.id);
-      if (!asset) {
-        continue;
-      }
-
-      const recalculate = asset.localDateTime !== _asset.localDateTime;
-      Object.assign(asset, _asset);
-
-      if (recalculate) {
-        assetsToRecalculate.push(asset);
-      }
-    }
-
-    this.removeAssets(assetsToRecalculate.map((asset) => asset.id));
-    this.addAssetsToBuckets(assetsToRecalculate);
+    const lookup = new Map<string, AssetResponseDto>(assets.map((asset) => [asset.id, asset]));
+    const { unprocessedIds } = this.#runAssetOperation(new Set(lookup.keys()), (asset) => {
+      updateObject(asset, lookup.get(asset.id));
+      return { remove: false };
+    });
+    return unprocessedIds.values().map((id) => lookup.get(id)!);
   }
 
   removeAssets(ids: string[]) {
-    const idSet = new Set(ids);
+    const { unprocessedIds } = this.#runAssetOperation(new Set(ids), () => {
+      return { remove: true };
+    });
+    return [...unprocessedIds];
+  }
 
-    // Iterate in reverse to allow array splicing.
-    for (let index = this.buckets.length - 1; index >= 0; index--) {
-      const bucket = this.buckets[index];
-      let changed = false;
-      for (let index_ = bucket.assets.length - 1; index_ >= 0; index_--) {
-        const asset = bucket.assets[index_];
-        if (!idSet.has(asset.id)) {
-          continue;
-        }
+  refreshLayout() {
+    for (const bucket of this.buckets) {
+      this.#updateGeometry(bucket, true);
+    }
+    this.updateIntersections();
+  }
 
-        bucket.assets.splice(index_, 1);
-        changed = true;
-        if (bucket.assets.length === 0) {
-          this.buckets.splice(index, 1);
-        }
+  getFirstAsset(): AssetResponseDto | undefined {
+    return this.buckets[0]?.getFirstAsset();
+  }
+
+  async getPreviousAsset(asset: AssetResponseDto): Promise<AssetResponseDto | undefined> {
+    let bucket = await this.#getBucketInfoForAsset(asset);
+    if (!bucket) {
+      return;
+    }
+
+    for (const group of bucket.dateGroups) {
+      const index = group.intersetingAssets.findIndex((ia) => ia.id === asset.id);
+      if (index > 0) {
+        return group.intersetingAssets[index - 1].asset;
       }
-      if (changed) {
-        bucket.dateGroups = splitBucketIntoDateGroups(bucket, get(locale));
-        this.updateGeometry(bucket, true);
+    }
+
+    let bucketIndex = this.buckets.indexOf(bucket) - 1;
+    while (bucketIndex >= 0) {
+      bucket = this.buckets[bucketIndex];
+      if (!bucket) {
+        return;
+      }
+      await this.loadBucket(bucket.bucketDate);
+      const previous = bucket.lastDateGroup?.intersetingAssets.at(-1)?.asset;
+      if (previous) {
+        return previous;
+      }
+      bucketIndex--;
+    }
+  }
+
+  async getNextAsset(asset: AssetResponseDto): Promise<AssetResponseDto | undefined> {
+    let bucket = await this.#getBucketInfoForAsset(asset);
+    if (!bucket) {
+      return;
+    }
+
+    for (const group of bucket.dateGroups) {
+      const index = group.intersetingAssets.findIndex((ia) => ia.id === asset.id);
+      if (index !== -1 && index < group.intersetingAssets.length - 1) {
+        return group.intersetingAssets[index + 1].asset;
       }
     }
+
+    let bucketIndex = this.buckets.indexOf(bucket) + 1;
+    while (bucketIndex < this.buckets.length - 1) {
+      bucket = this.buckets[bucketIndex];
+      await this.loadBucket(bucket.bucketDate);
+      const next = bucket.dateGroups[0]?.intersetingAssets[0]?.asset;
+      if (next) {
+        return next;
+      }
+      bucketIndex++;
+    }
   }
 
-  getFirstAsset(): AssetResponseDto | null {
-    if (this.buckets.length > 0 && this.buckets[0].assets.length > 0) {
-      return this.buckets[0].assets[0];
-    }
-    return null;
-  }
-
-  async getPreviousAsset(asset: AssetResponseDto): Promise<AssetResponseDto | null> {
-    const info = await this.getBucketInfoForAsset(asset);
-    if (!info) {
-      return null;
-    }
-
-    const { bucket, assetIndex, bucketIndex } = info;
-
-    if (assetIndex !== 0) {
-      return bucket.assets[assetIndex - 1];
-    }
-
-    if (bucketIndex === 0) {
-      return null;
-    }
-
-    const previousBucket = this.buckets[bucketIndex - 1];
-    await this.loadBucket(previousBucket.bucketDate);
-    return previousBucket.assets.at(-1) || null;
-  }
-
-  async getNextAsset(asset: AssetResponseDto): Promise<AssetResponseDto | null> {
-    const info = await this.getBucketInfoForAsset(asset);
-    if (!info) {
-      return null;
-    }
-
-    const { bucket, assetIndex, bucketIndex } = info;
-
-    if (assetIndex !== bucket.assets.length - 1) {
-      return bucket.assets[assetIndex + 1];
-    }
-
-    if (bucketIndex === this.buckets.length - 1) {
-      return null;
-    }
-
-    const nextBucket = this.buckets[bucketIndex + 1];
-    await this.loadBucket(nextBucket.bucketDate);
-    return nextBucket.assets[0] || null;
-  }
-
-  private isExcluded(asset: AssetResponseDto) {
+  isExcluded(asset: AssetResponseDto) {
     return (
-      isMismatched(this.options.isArchived ?? false, asset.isArchived) ||
-      isMismatched(this.options.isFavorite, asset.isFavorite) ||
-      isMismatched(this.options.isTrashed ?? false, asset.isTrashed)
+      isMismatched(this.#options.isArchived, asset.isArchived) ||
+      isMismatched(this.#options.isFavorite, asset.isFavorite) ||
+      isMismatched(this.#options.isTrashed, asset.isTrashed)
     );
   }
 }
