@@ -49,18 +49,21 @@ export function getJustifiedLayoutFromAssets(
 type Geometry = ReturnType<typeof createJustifiedLayout>;
 class Adapter {
   result;
+  width;
   constructor(result: Geometry) {
     this.result = result;
+    this.width = 0;
+    for (const box of this.result.boxes) {
+      if (box.top < 100) {
+        this.width = box.left + box.width;
+      } else {
+        break;
+      }
+    }
   }
 
   get containerWidth() {
-    let width = 0;
-    for (const box of this.result.boxes) {
-      if (box.top < 100) {
-        width = box.left + box.width;
-      }
-    }
-    return width;
+    return this.width;
   }
 
   get containerHeight() {
@@ -84,12 +87,6 @@ class Adapter {
   }
 }
 
-export const emptyGeometry = new Adapter({
-  containerHeight: 0,
-  widowCount: 0,
-  boxes: [],
-});
-
 export function justifiedLayout(assets: AssetResponseDto[], options: CommonLayoutOptions) {
   const adapter = {
     targetRowHeight: options.rowHeight,
@@ -103,4 +100,27 @@ export function justifiedLayout(assets: AssetResponseDto[], options: CommonLayou
     adapter,
   );
   return new Adapter(result);
+}
+
+export const emptyGeometry = () =>
+  new Adapter({
+    containerHeight: 0,
+    widowCount: 0,
+    boxes: [],
+  });
+
+export type CommonPosition = {
+  top: number;
+  left: number;
+  width: number;
+  height: number;
+};
+
+export function getPosition(geometry: CommonJustifiedLayout, boxIdx: number): CommonPosition {
+  const top = geometry.getTop(boxIdx);
+  const left = geometry.getLeft(boxIdx);
+  const width = geometry.getWidth(boxIdx);
+  const height = geometry.getHeight(boxIdx);
+
+  return { top, left, width, height };
 }
