@@ -456,10 +456,10 @@
 </UserPageLayout>
 
 {#if selectHidden}
-  <div
+  <dialog
+    open
     transition:fly={{ y: innerHeight, duration: 150, easing: quintOut, opacity: 0 }}
     class="absolute left-0 top-0 z-[9999] h-full w-full bg-immich-bg dark:bg-immich-dark-bg"
-    role="dialog"
     aria-modal="true"
     aria-labelledby="manage-visibility-title"
     use:focusTrap
@@ -471,5 +471,5 @@
       onClose={() => (selectHidden = false)}
       {loadNextPage}
     />
-  </div>
+  </dialog>
 {/if}
