@@ -24,6 +24,7 @@
   import { mdiPencil, mdiPlus, mdiTag, mdiTagMultiple, mdiTrashCanOutline } from '@mdi/js';
   import { t } from 'svelte-i18n';
   import type { PageData } from './$types';
+  import { onDestroy } from 'svelte';
 
   interface Props {
     data: PageData;
@@ -39,8 +40,9 @@
   const buildMap = (tags: TagResponseDto[]) => {
     return Object.fromEntries(tags.map((tag) => [tag.value, tag]));
   };
-
-  const assetStore = new AssetStore({});
+  const assetStore = new AssetStore();
+  $effect(() => void assetStore.updateOptions({ deferInit: !tag, tagId }));
+  onDestroy(() => assetStore.destroy());
 
   let tags = $state<TagResponseDto[]>([]);
   $effect(() => {
@@ -51,10 +53,6 @@
   let tag = $derived(currentPath ? tagsMap[currentPath] : null);
   let tagId = $derived(tag?.id);
   let tree = $derived(buildTree(tags.map((tag) => tag.value)));
-
-  $effect.pre(() => {
-    void assetStore.updateOptions({ tagId });
-  });
 
   const handleNavigation = async (tag: string) => {
     await navigateToView(normalizeTreePath(`${data.path || ''}/${tag}`));
